fix(webui): fix polling & add timepicker to flow graph
