+++ webui/src/pages/config/components/AddDialog/index.module.css
@@ -2,6 +2,7 @@
     display: flex;
     align-items: center;
     margin-bottom: 20px;
+
     span {
         margin-right: 10px;
     }
